+++ xml_schema/src/xsd/attribute_group.rs
