+++ xml_schema/src/xsd/complex_content.rs
@@ -1,0 +1,7 @@
+use crate::xsd::extension::Extension;
+
+#[derive(Clone, Default, Debug, PartialEq, YaDeserialize)]
+#[yaserde(prefix = "xs", namespace = "xs: http://www.w3.org/2001/XMLSchema")]
+pub struct ComplexContent {
+  pub extension: Option<Extension>,
+}
