+++ xml_schema/src/xsd/complex_type.rs
@@ -1,0 +1,24 @@
+use crate::xsd::{
+  annotation::Annotation, attribute::Attribute, complex_content::ComplexContent,
+  sequence::Sequence, simple_content::SimpleContent,
+};
+
+#[derive(Clone, Default, Debug, PartialEq, YaDeserialize)]
+#[yaserde(
+  rename = "complexType"
+  prefix = "xs",
+  namespace = "xs: http://www.w3.org/2001/XMLSchema"
+)]
+pub struct ComplexType {
+  #[yaserde(attribute)]
+  pub name: String,
+  #[yaserde(rename = "attribute")]
+  pub attributes: Vec<Attribute>,
+  pub sequence: Option<Sequence>,
+  #[yaserde(rename = "simpleContent")]
+  pub simple_content: Option<SimpleContent>,
+  #[yaserde(rename = "complexContent")]
+  pub complex_content: Option<ComplexContent>,
+  #[yaserde(rename = "annotation")]
+  pub annotation: Option<Annotation>,
+}
