+++ xml_schema/src/xsd/element.rs
@@ -1,0 +1,37 @@
+use crate::xsd::{
+  annotation::Annotation, complex_type::ComplexType, max_occurences::MaxOccurences,
+  simple_type::SimpleType,
+};
+
+#[derive(Clone, Default, Debug, PartialEq, YaDeserialize)]
+#[yaserde(prefix = "xs", namespace = "xs: http://www.w3.org/2001/XMLSchema")]
+pub struct Element {
+  #[yaserde(attribute)]
+  pub name: String,
+  #[yaserde(rename = "type", attribute)]
+  pub kind: Option<String>,
+  #[yaserde(rename = "ref", attribute)]
+  pub refers: Option<String>,
+  #[yaserde(rename = "minOccurs", attribute)]
+  pub min_occurences: Option<u64>,
+  #[yaserde(rename = "maxOccurs", attribute)]
+  pub max_occurences: Option<MaxOccurences>,
+  #[yaserde(rename = "complexType")]
+  pub complex_type: Option<ComplexType>,
+  #[yaserde(rename = "simpleType")]
+  pub simple_type: Option<SimpleType>,
+  #[yaserde(rename = "annotation")]
+  pub annotation: Option<Annotation>,
+}
+
+impl Element {
+  pub fn get_refers(&self) -> Option<&str> {
+    self.refers.as_ref().and_then(|refers| {
+      if refers.is_empty() {
+        None
+      } else {
+        Some(refers.as_str())
+      }
+    })
+  }
+}
