+++ xml_schema/src/xsd/extension.rs
@@ -1,0 +1,18 @@
+use crate::xsd::{attribute::Attribute, group::Group, sequence::Sequence};
+
+#[derive(Clone, Default, Debug, PartialEq, YaDeserialize)]
+#[yaserde(
+  root = "extension",
+  prefix = "xs",
+  namespace = "xs: http://www.w3.org/2001/XMLSchema"
+)]
+pub struct Extension {
+  #[yaserde(attribute)]
+  pub base: String,
+  #[yaserde(rename = "attribute")]
+  pub attributes: Vec<Attribute>,
+  #[yaserde(rename = "sequence")]
+  pub sequences: Vec<Sequence>,
+  #[yaserde(rename = "group")]
+  pub group: Option<Group>,
+}
