+++ xml_schema/src/xsd/list.rs
@@ -1,0 +1,6 @@
+#[derive(Clone, Default, Debug, PartialEq, YaDeserialize)]
+#[yaserde(prefix = "xs", namespace = "xs: http://www.w3.org/2001/XMLSchema")]
+pub struct List {
+  #[yaserde(rename = "itemType", attribute)]
+  pub item_type: String,
+}
