+++ xml_schema/src/xsd/qualification.rs
