+++ xml_schema/src/xsd/restriction.rs
@@ -1,0 +1,6 @@
+#[derive(Clone, Default, Debug, PartialEq, YaDeserialize)]
+#[yaserde(prefix = "xs", namespace = "xs: http://www.w3.org/2001/XMLSchema")]
+pub struct Restriction {
+  #[yaserde(rename = "base", attribute)]
+  pub base: Option<String>,
+}
