+++ xml_schema_derive/src/xsd/annotation.rs
@@ -1,24 +1,7 @@
-use crate::xsd::{attribute::Attribute, Implementation, XsdContext};
+use crate::xsd::{Implementation, XsdContext};
 use proc_macro2::TokenStream;
 
-#[derive(Clone, Default, Debug, PartialEq, YaDeserialize)]
-#[yaserde(
-    rename = "annotation"
-    prefix = "xs",
-    namespace = "xs: http://www.w3.org/2001/XMLSchema"
-  )]
-pub struct Annotation {
-  #[yaserde(attribute)]
-  pub id: Option<String>,
-  #[yaserde(rename = "attribute")]
-  pub attributes: Vec<Attribute>,
-  #[yaserde(
-      rename = "documentation"
-      prefix = "xs",
-      namespace = "xs: http://www.w3.org/2001/XMLSchema"
-    )]
-  pub documentation: Vec<String>,
-}
+use xml_schema::Annotation;
 
 impl Implementation for Annotation {
   fn implement(
