+++ xml_schema_derive/src/xsd/attribute.rs
@@ -1,41 +1,9 @@
-use crate::xsd::{
-  rust_types_mapping::RustTypesMapping, simple_type::SimpleType, Implementation, XsdContext,
-};
+use crate::xsd::{rust_types_mapping::RustTypesMapping, Implementation, XsdContext};
 use heck::ToSnakeCase;
 use proc_macro2::{Span, TokenStream};
 use syn::Ident;
 
-#[derive(Clone, Default, Debug, PartialEq, YaDeserialize)]
-#[yaserde(
-  rename = "attribute",
-  prefix = "xs",
-  namespace = "xs: http://www.w3.org/2001/XMLSchema"
-)]
-pub struct Attribute {
-  #[yaserde(prefix = "xs", attribute)]
-  pub name: Option<String>,
-  #[yaserde(rename = "type", attribute)]
-  pub kind: Option<String>,
-  // #[yaserde(attribute)]
-  // pub default: Option<String>,
-  // #[yaserde(attribute)]
-  // pub fixed: Option<String>,
-  #[yaserde(rename = "use", attribute)]
-  pub required: Required,
-  #[yaserde(rename = "ref", attribute)]
-  pub reference: Option<String>,
-  #[yaserde(rename = "simpleType")]
-  pub simple_type: Option<SimpleType>,
-}
-
-#[derive(Clone, Debug, Default, PartialEq, YaDeserialize)]
-pub enum Required {
-  #[default]
-  #[yaserde(rename = "optional")]
-  Optional,
-  #[yaserde(rename = "required")]
-  Required,
-}
+use xml_schema::{Attribute, Required};
 
 impl Implementation for Attribute {
   fn implement(
