+++ xml_schema_derive/src/xsd/complex_content.rs
@@ -1,18 +1,10 @@
-use crate::xsd::{extension::Extension, xsd_context::XsdContext};
+use crate::xsd::Implementation;
 use proc_macro2::TokenStream;
 
-#[derive(Clone, Default, Debug, PartialEq, YaDeserialize)]
-#[yaserde(prefix = "xs", namespace = "xs: http://www.w3.org/2001/XMLSchema")]
-pub struct ComplexContent {
-  pub extension: Option<Extension>,
-}
+use xml_schema::{ComplexContent, XsdContext};
 
-impl ComplexContent {
-  pub fn get_field_implementation(
-    &self,
-    context: &XsdContext,
-    prefix: &Option<String>,
-  ) -> TokenStream {
+impl Implementation for ComplexContent {
+  fn get_field_implementation(&self, context: &XsdContext, prefix: &Option<String>) -> TokenStream {
     self
       .extension
       .as_ref()
