+++ xml_schema_derive/src/xsd/list.rs
@@ -1,12 +1,7 @@
 use crate::xsd::{rust_types_mapping::RustTypesMapping, Implementation, XsdContext};
 use proc_macro2::{Ident, TokenStream};
 
-#[derive(Clone, Default, Debug, PartialEq, YaDeserialize)]
-#[yaserde(prefix = "xs", namespace = "xs: http://www.w3.org/2001/XMLSchema")]
-pub struct List {
-  #[yaserde(rename = "itemType", attribute)]
-  pub item_type: String,
-}
+use xml_schema::List;
 
 impl Implementation for List {
   fn implement_childs(
