+++ xml_schema_derive/src/xsd/sequence.rs
@@ -1,13 +1,8 @@
-use crate::xsd::{element::Element, Implementation, XsdContext};
+use crate::xsd::{Implementation, XsdContext};
 use log::info;
 use proc_macro2::TokenStream;
 
-#[derive(Clone, Default, Debug, PartialEq, YaDeserialize)]
-#[yaserde(prefix = "xs", namespace = "xs: http://www.w3.org/2001/XMLSchema")]
-pub struct Sequence {
-  #[yaserde(rename = "element")]
-  pub elements: Vec<Element>,
-}
+use xml_schema::Sequence;
 
 impl Implementation for Sequence {
   fn implement(
@@ -23,10 +18,8 @@ impl Implementation for Sequence {
       .map(|element| element.get_field_implementation(context, prefix))
       .collect()
   }
-}
 
-impl Sequence {
-  pub fn get_sub_types_implementation(
+  fn get_sub_types_implementation(
     &self,
     context: &XsdContext,
     namespace_definition: &TokenStream,
@@ -40,11 +33,7 @@ impl Sequence {
       .collect()
   }
 
-  pub fn get_field_implementation(
-    &self,
-    context: &XsdContext,
-    prefix: &Option<String>,
-  ) -> TokenStream {
+  fn get_field_implementation(&self, context: &XsdContext, prefix: &Option<String>) -> TokenStream {
     self
       .elements
       .iter()
