+++ xml_schema_derive/src/xsd/simple_content.rs
@@ -1,12 +1,7 @@
-use crate::xsd::{extension::Extension, Implementation, XsdContext};
+use crate::xsd::{Implementation, XsdContext};
 use proc_macro2::TokenStream;
 
-#[derive(Clone, Default, Debug, PartialEq, YaDeserialize)]
-#[yaserde(prefix = "xs", namespace = "xs: http://www.w3.org/2001/XMLSchema")]
-pub struct SimpleContent {
-  #[yaserde(prefix = "xs", rename = "extension")]
-  pub extension: Extension,
-}
+use xml_schema::SimpleContent;
 
 impl Implementation for SimpleContent {
   fn implement(
@@ -19,14 +14,8 @@ impl Implementation for SimpleContent {
       .extension
       .implement(namespace_definition, prefix, context)
   }
-}
 
-impl SimpleContent {
-  pub fn get_field_implementation(
-    &self,
-    context: &XsdContext,
-    prefix: &Option<String>,
-  ) -> TokenStream {
+  fn get_field_implementation(&self, context: &XsdContext, prefix: &Option<String>) -> TokenStream {
     self.extension.get_field_implementation(context, prefix)
   }
 }
