+++ xml_schema_derive/tests/complex_type.rs
@@ -5,7 +5,7 @@ use yaserde::ser::to_string;
 #[test]
 fn complex_type_string() {
   #[derive(Debug, XmlSchema)]
-  #[xml_schema(source = "xml_schema/tests/complex_type.xsd")]
+  #[xml_schema(source = "xml_schema_derive/tests/complex_type.xsd")]
   struct ComplexTypeSchema;
 
   let xml_1 = r#"
