+++ xml_schema_derive/tests/simple_type.rs
@@ -2,11 +2,12 @@ use xml_schema_derive::XmlSchema;
 use yaserde::de::from_str;
 use yaserde::ser::to_string;
 
+
 #[test]
 fn simple_type_string() {
   #[derive(Debug, XmlSchema)]
   #[xml_schema(
-    source = "xml_schema/tests/simple_type_string.xsd",
+    source = "xml_schema_derive/tests/simple_type_string.xsd",
     target_prefix = "st"
   )]
   struct SimpleTypeSchema;
@@ -29,7 +30,7 @@ fn simple_type_string() {
 #[test]
 fn simple_type_list() {
   #[derive(Debug, XmlSchema)]
-  #[xml_schema(source = "xml_schema/tests/simple_type_list.xsd")]
+  #[xml_schema(source = "xml_schema_derive/tests/simple_type_list.xsd")]
   struct SimpleTypeSchema;
 
   let xml_1 = r#"
