+++ xml_schema_derive/tests/simple_type_list.xsd
